+++ src/Filters/CertificatesSliderXxl.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Cher4geo35\Sliders\Filters;
+
+use Intervention\Image\Facades\Image;
+use Intervention\Image\Filters\FilterInterface;
+use Intervention\Image\Image as File;
+
+class CertificatesSliderXxl implements FilterInterface {
+    public function applyFilter(File $image)
+    {
+        $image->orientate();
+        return $image
+            ->widen(245);
+    }
+}